+++ src/ex_eval.c
@@ -251,7 +251,7 @@ cause_errthrow(
 	    while (*plist != NULL)
 		plist = &(*plist)->next;
 
-	    elem = (struct msglist *)alloc((unsigned)sizeof(struct msglist));
+	    elem = (struct msglist *)alloc(sizeof(struct msglist));
 	    if (elem == NULL)
 	    {
 		suppress_errthrow = TRUE;
@@ -519,7 +519,7 @@ throw_exception(void *value, except_type_T type, char_u *cmdname)
 	}
     }
 
-    excp = (except_T *)alloc((unsigned)sizeof(except_T));
+    excp = (except_T *)alloc(sizeof(except_T));
     if (excp == NULL)
 	goto nomem;
 
@@ -1441,7 +1441,7 @@ ex_try(exarg_T *eap)
 	    {
 		eslist_T	*elem;
 
-		elem = (eslist_T *)alloc((unsigned)sizeof(struct eslist_elem));
+		elem = (eslist_T *)alloc(sizeof(struct eslist_elem));
 		if (elem == NULL)
 		    emsg(_(e_outofmem));
 		else
